+++ transactions-api/V1/Factories/AbstractTransactionFactory.cs
@@ -1,0 +1,16 @@
+using System.Collections.Generic;
+using System.Linq;
+using transactions_api.V1.Domain;
+
+namespace transactions_api.V1.Factory
+{
+    public abstract class AbstractTransactionFactory
+    {
+        public abstract Transaction FromUhTransaction(UhTransaction transaction);
+
+        public List<Transaction> FromUhTransaction(IEnumerable<UhTransaction> result)
+        {
+            return result.Select(FromUhTransaction).ToList();
+        }
+    }
+}
