+++ transactions-api/V1/Factories/TransactionFactory.cs
@@ -1,0 +1,17 @@
+using transactions_api.V1.Domain;
+
+namespace transactions_api.V1.Factory
+{
+    public class TransactionFactory : AbstractTransactionFactory
+    {
+        public override Transaction FromUhTransaction(UhTransaction transaction)
+        {
+            return new Transaction
+            {
+                Balance = transaction.Balance,
+                Code = transaction.Code,
+                Date = transaction.Date
+            };
+        }
+    }
+}
