+++ transactions-api/V1/Gateways/TransactionsGateway.cs
@@ -2,6 +2,7 @@
 using System.Collections.Generic;
 using System.Linq;
 using transactions_api.V1.Domain;
+using transactions_api.V1.Factory;
 using UnitTests.V1.Infrastructure;
 
 namespace UnitTests.V1.Gateways
@@ -9,22 +10,18 @@ namespace UnitTests.V1.Gateways
     public class TransactionsGateway : ITransactionsGateway
     {
         private readonly IUHContext _uhcontext;
+        private readonly TransactionFactory _transactionFactory;
 
         public TransactionsGateway(IUHContext uhcontext)
         {
+            _transactionFactory = new TransactionFactory();
             _uhcontext = uhcontext;
         }
 
         public List<Transaction> GetTransactionsByPropertyRef(string propertyRef)
         {
             var result = _uhcontext.UTransactions.Where(t => t.PropRef == propertyRef).ToList();
-
-            return ConvertToTransactions(result);
-        }
-
-        private static List<Transaction> ConvertToTransactions(IEnumerable<UhTransaction> result)
-        {
-            return result.Select(Transaction.fromUHTransaction).ToList();
+            return _transactionFactory.FromUhTransaction(result);
         }
     }
 }
